chore(app): introduce PDD and `./app/components` directory
